Move `tokenFactoryReset` to EZSPv13, it's not in v8
